Update debian/ubuntu versions to use a single `distro` variable
Prior to this commit, two variables (allDistros, serverDistros)
existed for defining the distros on which certain binaries should
be built. As there is no delineation between "server" and "non-server"
distros for the binaries we build, there is no need for separate
variables to describe them. This commit collapses the variables
into a single `distro` variable.

Signed-off-by: Stephen Augustus <saugustus@vmware.com>