Update symlinks for all debian-like package definitions
We currently use a single build definition (for bionic) and symlink
folders for each debian-like distro to it. Previously, several of these
build definitions only contained a symlink for the kubectl binary.
This commit updates the symlinks so that all kubernetes binaries are
built for all debian-like distros that we have defined.

Signed-off-by: Stephen Augustus <saugustus@vmware.com>