rpms: change CNI version check to ">="
Pull https://github.com/kubernetes/release/pull/716 changed the Deb
package metadata to check the CNI version with a ">=".

There is still ongoing debate on the right way to curate our package
metadata relative to dependencies.  But in the meantime...We need this
to be consistent across packaging.

So this commit does the same as #716 but also for RPMs.

Signed-off-by: Tim Pepper <tpepper@vmware.com>